API/requirements.txt
pandas ==2.0.3
python-dotenv==1.0.1
openai==1.50.2
mlxtend==0.23.1
firebase-admin==7.1.0
google-cloud-storage==2.18.2
pinecone
gradio